+++ pyzoo/test/zoo/automl/model/test_Seq2Seq.py
@@ -21,7 +21,7 @@ import pytest
 
 from test.zoo.pipeline.utils.test_utils import ZooTestCase
 from zoo.automl.model.Seq2Seq import LSTMSeq2Seq
-from zoo.automl.feature.time_sequence import TimeSequenceFeatureTransformer
+from zoo.zouwu.feature.time_sequence import TimeSequenceFeatureTransformer
 from numpy.testing import assert_array_almost_equal
 import pandas as pd
 import numpy as np
