+++ pyzoo/test/zoo/zouwu/model/forecast/test_mtnet_forecaster.py
@@ -16,7 +16,7 @@
 
 import pytest
 import numpy as np
-from zoo.automl.feature.time_sequence import TimeSequenceFeatureTransformer
+from zoo.zouwu.feature.time_sequence import TimeSequenceFeatureTransformer
 import tensorflow as tf
 import pandas as pd
