+++ pyzoo/zoo/automl/pipeline/time_sequence.py
@@ -18,7 +18,7 @@ import time
 from zoo.automl.common.metrics import Evaluator
 from zoo.automl.pipeline.abstract import Pipeline
 from zoo.automl.common.util import *
-from zoo.automl.feature.time_sequence import TimeSequenceFeatureTransformer
+from zoo.zouwu.feature.time_sequence import TimeSequenceFeatureTransformer
 from zoo.automl.model.time_sequence import TimeSequenceModel
 from zoo.automl.common.parameters import *
 
@@ -220,7 +220,7 @@ def load_ts_pipeline(file):
 
 
 def load_xgboost_pipeline(file, model_type="regressor"):
-    from zoo.automl.feature.identity_transformer import IdentityTransformer
+    from zoo.zouwu.feature.identity_transformer import IdentityTransformer
     from zoo.automl.model import XGBoost
     feature_transformers = IdentityTransformer()
     model = XGBoost(model_type=model_type)
